Cleanup: Code quality improvements and optimizations
- Box P2PSession in SessionInner enum to reduce variant size (~440 bytes → ~228 bytes)
- Use iterator methods instead of manual loops (find(), map())
- Use range contains checks instead of manual comparisons
- Reduce DrawCommand visibility to pub(crate)
- Remove unnecessary type casts
- Use array literals instead of vec! where applicable

+++ core/src/integration.rs
@@ -1118,7 +1118,7 @@ mod tests {
             game.update(1.0 / 60.0).unwrap();
             game.render().unwrap();
             if frame == 3 {
-                frame3_snapshot = Some(state_manager.save_state(&mut game, frame as i32).unwrap());
+                frame3_snapshot = Some(state_manager.save_state(&mut game, frame).unwrap());
             }
         }
